requirements.txt
facebook-sdk==3.1.0
python-dotenv==1.0.0
schedule==1.2.0
Pillow==9.5.0
playwright
requests==2.31.0
anthropic==0.8.1 